+++ CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Correct histogram anomaly checking
 - Avoid errors with very small `multipleOf` values
 - Avoid errors with `multipleOf` and very different scales
+- Don't anchor string matches in `StaticPatternProperty`
 
 ## [0.15.0]
 ### Changed